Added center stage icon. Refactor center stage code. Add AV capture util.
