+++ Signal/Images.xcassets/centre-stage-28.imageset/Contents.json
@@ -1,17 +1,8 @@
 {
   "images" : [
     {
-      "filename" : "portrait-mode copy.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "filename" : "portrait-mode8.pdf",
+      "idiom" : "universal"
     }
   ],
   "info" : {
